tts-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
motor==3.3.1
pydub==0.25.1
firebase-admin==6.2.0
PyPDF2==3.0.1
python-docx==1.0.1
beautifulsoup4==4.8.0
extract-msg==0.23.1
langdetect==1.0.9
nltk==3.8.1
chardet==3.0.4
textract==1.6.3
PyYAML==6.0.1
aiofiles==23.2.1
ffmpeg-python==0.2.0
jinja2==3.1.2
rich==13.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
Pillow==10.1.0
python-slugify==8.0.1
pandas==2.1.2
numpy==1.26.1
tensorflow==2.14.0
soundfile==0.12.1
librosa==0.10.1